Specify port on the placeholder service.
While somewhat superfluous in general, this indicates what port(s) the placeholder service expects to route traffic on.  This is just `80` because cluster-locally we don't to auto-TLS.

This also serves the role of specifying the application protocol according to our port naming convention, to enable composition.  We use http/2 because all conformant gateways must support it, and it can be downgraded to http/1, but nothe the other way around.