Do not overwrite Kourier yaml when update namespace
When namespace in Kourier is updated, the yaml is overwritten. It
works a single test, but e2e-tests.sh run two tests e2e and auto-tls
so the second test always fails due to namespace not found error.

This patch fixes it by creating a new yaml file in tmp directory.