Add initial PDB for activator.
I went with conservative not more than 20% of tasks out at the same time.
Activators restart on my cluster pretty fast and the state is mostly in
relation to the requests in progress through _this_ activator,
so this should not make the rollout much longer, or disrupt it by a lot
even in the face of subseting.

Obviously once the PDB is in we can tweak it in either direction (as far
as not more than 1 unavailable at a time).