Create unix domain socket file in temp directory (#9211)
* Create unix domain socket file in temp directory

Currently queue-proxy creates unix domain socket file in the "current"
directory, where is /ko-app/ in upstream.

It works when root user runs queue-proxy or when the directory has
runtime user's permission. In other words, `/ko-app/` dir must have
runtime user's permission when nonroot users run queue-proxy.

It is a pain that the directory permission must be cared.

To solve the permission problems, this patch changes to create the
socket file under temp directory.

* Use filepath.Join