Remove JSON tag of DeprecatedGeneration in v1alpha1.Route. (#9231)
`go vet` barks at this and we're fairly sure this is safe to drop as v1alpha1 isn't used as a storage version anymore anyway and newer API versions do not have this field at all anymore. Nothing in our code keys off of it either, so this is the least invasive change to make go vet happy without breaking the golang struct.