Fix intermittent 404 in TestServiceWithTrafficSplit
While debugging the linked issue, I noticed that the test was mistakenly thinking that the Service was Ready before the kingress had even been reprogrammed (let alone probed!).  When I drilled into why I discovered that it is possible for the Route controller, when it hits an error creating a placeholder service to trigger a status update that inadvertently reports the Route as Ready before it has even updated the kingress resource.

The sequence of events here is:
1. `PreProcessReconcile` changes `Ready: Unknown` in preparation for an `ObservedGeneration` bump.
2. The early traffic assignment computation returns the condition to `Ready: True`.
3. Before we even reach the `kingress` update, the placeholder service update errors out.
4. `PostProcessReconcile` bumps the `ObservedGeneration`.
5. `UpdateStatus` records the status.

The net effect of this is that we have put the `Route` into an `IsReady()` state before we've even attempted to reconcile the `kingress`, prompting the tests to probe and get a 404.

The fix is to `MarkIngressNotConfigured` upon observing new generations, since this is the last thing we'll flip back on our path to `Ready: True`, but I wonder whether `PreProcessReconcile` should take a callback to let (require?) the reconciler itself to specify how to properly adjust the conditions.

Fixes: https://github.com/knative-sandbox/net-contour/issues/238