Change the way worker timeout is handled. (#9389)
The way the "worker" timeout is handled today is effectively a global deadline for ALL ksvcs, which feels inconsistent with the comment above the value.  As things stand today, we have 5 minutes from when the test starts to create all N services, and we only create 50 at a time, so with scale-200, 4 waves have 5 minutes to complete (1m15s is a *very* reasonable MTTR under stress).  The current sense of this parameter doesn't really provide much value over the `-timeout` flag to `go test` IMO.

This changes the way this duration is evaluated to start the timer when the ksvc is created, and *lowers* the bound itself such that each ksvc has 3 minutes to become ready after it is created.  The `-timeout` will still act as a global bound, but this will give each ksvc a more reasonable bound.  I suspect we can pull this value down over time, but I'm starting somewhat conservatively.