Do not use ResolvableDomain for cluster local domain
`helloworldURL` is cluster local domain (`svc.cluster.locl`) when `accessibleExternal` is false.
In this case, the ResolvableDomain option is not available as the host name cannot be resolved from outside cluster.

Here is the sample log when ResolvableDomain is enabled:

```
    spoof.go:119: Spoofing svc-to-svc-via-activ-dshkjqoi.serving-tests.svc.cluster.local -> svc-to-svc-via-activ-dshkjqoi.serving-tests.svc.cluster.local
    spoof.go:189: Retrying http://svc-to-svc-via-activ-dshkjqoi.serving-tests.svc.cluster.local: retrying for DNS error: Get "http://svc-to-svc-via-activ-dshkjqoi.serving-tests.svc.cluster.local": dial tcp: lookup svc-to-svc-via-activ-dshkjqoi.serving-tests.svc.cluster.local: no such host
```

This patch changes to stop using `ResolvableDomain` when `accessibleExternal` is false.