Run e2e tests on Github Actions (#9445)
* Run e2e tests on Github Actions

This change adds a Github Action to run the e2e and conformance tests on KinD.  The tests fan out a matrix with the dimensions `{k8s version} x {test package}` and this parallelism actually makes things quite fast.

One of the key motivations of this change is to leverage KinD to start getting proper coverage of the K8s versions that we claim to "support" in advance of releases.  Right now we are only testing 1.16.x on GKE, but this PR adds coverage of 1.17.x, 1.18.x, and 1.19.x.  Given that we're now running these tests 3x it felt wasteful to make them too regular, so this is set up so that we can run a different kingress implementation with each K8s version.  Right now this is set up with Kourier and Contour while we get Istio working with NodePort.

This doesn't run the scale test (b/c KinD) or any of the feature flag tests (b/c requires patched installs), but this ensures we have fairly solid coverage of the core elements of our test suite across the supported windo of K8s versions.

* use matrix.kingress in common patch

* Fix comment on JWT fix

* Fix the variable name :facepalm:

* Remove on-push trigger

* Use 0.9.0 everywhere and vary node versions