Set the default in the deploy.go and propagate config to build pod spec
This is the next step for #8563.

Here we set the EnableServiceLink to the default value set in the CM.
Right now it will match 1 to 1 to what is set on the revision.
But in future, when we change the default (in the next PR) this is
going to alter the existing deployments, keeping the
service/config/revision spec intact.

This also required changing dry run code, but it is done in webhook
and does not really rely on configmaps, so a nil is passed there.

Change-Id: Ib48dd6f93db44e0a0af9a94500ed542f371f3284