Add logging to aid with debugging (#9698)
* Add logging to aid with debugging

The issue with a catchy number #7727 shows that something very fishy is going on.
Basically the logs show that we did not apply the `minScale` and instead
applied the desired scale from the autoscaler (want=2).

This should help when this flakes again (which seems quite often) to aid
with debugging.

Change-Id: Ic42c455893273b1062fd9eaa6bf3aac80d006602

* comma

Change-Id: I48ca89da49af2f38393773998b5539a196f9311e