Simplify the steps to reconcile ingress a bit (#10120)
* Simplify the steps to reconcile ingress a bit

In order to optimize the rollout computation it is preferential that `reconcileIngress` method itself has the access to the ingress object creation, rather than being passed in.
This way (in the next PR) depending on the `Get` result we can either
create new ingress with affixed rollout or create new one with already
_rolled_ one — which means we won't have to recompute the paths/splits,
but rather we'd just create them according to the rollout from the
beginning.

Change-Id: I074dbd3e04eab8e68fb7537faa4faa644f4fc655

* style

Change-Id: I5d30a6115c0e00569460ccc3ad2c15df6869194c