Enabling reusing probe connection in activator
When a pod starts up, the activator probes it with the "K-Network-Probe"
header to establish readiness before proxying the actual request.
Since the probe (before this change) did not share a transport with
the proxy, the probe connection couldn't be reused, meaning we pointlessly
(and quite expensively) connected twice on pod startup. Sharing the
transport allows the probe connection to be reused by the actual
requests, avoiding the extra connect overhead.