Fix the bug when the rollout was in a busy loop preventing ingress from being ready
The bug was introduced, thanks for @markusthoemmes for finding it, when
I added `stepRevisions` call.
That call would compute the next rollout step time and update the
corresponding field on the rollout.
Unfortunately this can happen before `ObserveReady` is called and the
rest of the rollout params are computed.
This lead to `step=0`, which meant we'd never progress and constantly
update the next step time, creating a busy loop of sorts.

The fix is to not step revisions until we have computed the rest of the
stats, basically by gating `stepSize > 0`, which is guaranteed to be at
least 1, when the readiness is observed.