Add a functional component test to verify proper reconcile of the rollout (#10348)
* Move observe before step

Move observe to apply to prev before step.
Why? Because:

- otherwise we'll need observe to duplicate the same logic to compute minimal returned next step time,
  right now it doesn't and that is a bug

Is it safe?
- Yes it is. Since it only applies to existing rollouts. If an existing rollout is being `obeserved` and then a new revision is added, the stats will just be reset at the `Step` call. If `prev` is `nil` — check is added for that.

That required to change the table tests to setup the `enqueueAfter`
since now that is triggered by the table tests.

Change-Id: I253254085bba1e1016dfd211e01ab71a079faa5f

* Add a functional component test to verify proper concile

Add a multi-step component integration tests for reconcileIngress to ensure
proper rollout procedures are followed.
This is hard to do within standard table test.

Change-Id: I5fe409d1feea6ba972571843983a5632d2914a68

* nit

Change-Id: Ia706010a438284106ef88b8bd264de0fcd1b21f7