Compute and set the route status to match rollout status
Basically apply similar logic when computing the splits in the ingress,
but slightly simpler.

If there's a matching rollout, then pass its revisions to the code that
existed already for a single target (now it's a loop -- but otherwise
basically the same).

For tagged routes during rollout it will display the split over the
tagged route (it feels more natural that way).

Change-Id: I6912f563b15fcf28d5d3fd926f32e2669f4a7105