Compute and set the Route.Status to match rollout status (#10420)
* Compute and set the route status to match rollout status

Basically apply similar logic when computing the splits in the ingress,
but slightly simpler.

If there's a matching rollout, then pass its revisions to the code that
existed already for a single target (now it's a loop -- but otherwise
basically the same).

For tagged routes during rollout it will display the split over the
tagged route (it feels more natural that way).

Change-Id: I6912f563b15fcf28d5d3fd926f32e2669f4a7105

* some shuffles

Change-Id: I912527e46e47c2d42ef3c9339bafb57f58a4fe15

* fix-aliasing

Change-Id: I01bddda88c6953115e1b540405cad52046df7a04