Take 1% time into account (#10449)
* Take 1% time into account

When computing rollout properties, take into account the duration
of the first step. Especially, if the first step is noticeably large
compared to the overall rollout duration, or even larger than it.
Before we'd rollout in possibly much longer than duration time,
with this, we'd be much closer to that time.

* fixtest