Simplify and optimise proxy header pruning
Adds a util.NewHeaderPruningReverseProxy method directly returning a
ReverseProxy that strips activator headers.

This avoids needing util.SetupHeaderPruning which was pulling out the
inner Director func from NewSingleHostReverseProxy and wrapping a
closure around it (the Director func we need to write instead is super
trivial since we don't need base paths/query strings so we were calling
NewSingleHostReverseProxy and introducing a closure just to save writing
four or so lines before).

Additionally since we never want to proxy at a different base path or
scheme or with extra query string params we can accept a string
targetHost here which avoids having to create a garbage url.URL on each
request in throttler.

Doing this removes a little bit of duplication since the same code
needed to be in both QP and activator.  Also it's a bit faster, saves a
couple allocations on every request and (I think) it's clearer.