Use explicit revID parameter in Throttler (#10545)
The revisionID is fundamental to what Throttler is doing, so better to
be explicit and not smuggle it in in context, especially since Throttler
is not a handler.

This avoids having to do util.WithRevID(ctx) stuff in
the tests, but more importantly means we can (in a follow-on PR) move
WithRevID and RevIDFrom back in to `handler` and make them private,
setting us up to get rid of the `util` package.