Set useragent for digest resolver
This takes advantage of a somewhat recent addition to ggcr that added
remote.WithUserAgent as an option. Setting this string will wrap the
User-Agent header with additional go-containerregistry version
information, when present, which is a nice thing to do as a citizen of
the internet.

Prior to this, we would send:

User-Agent: go-containerregistry[/${VERSION}]

Now we will send:

User-Agent: knative[/${COMMIT}] (serving) go-containerregistry[/${VERSION}]

Note that sometimes version information is not available. In some cases,
we do our best and just send a string without version information.