Rever the 100ms to 1ms change, since it flakes a lot (#10609)
This basically expects the test server to reply within 1ms
which is not possible in many cases, especially on prow, thus the
variable `probed` is never updated, since the handler is terminated
before that is invoked.

This should not matter anyway, since all the other test cases use 10s
anyway (which is the default).

A good set is here: https://prow.knative.dev/pr-history/?org=knative&repo=serving&pr=10604