Do not pass context to the scale, but logger instead
Right now Scale accepts context, but it is only used to acquire logger.
While not an issue in itself (besides random performance losses going
through the linkedlist that is context values), the way context is
passed _is_ a problem though.

While trying to sort out #10642 the obvious thing to look is that we use
reconcile context in a some long running process. _This_ is one of
places. And while this probably *is not* the problem, it is not good to
use the context beyond its lifetime.

There are a few ways to solve it. I went with a hammer and just stored
the logger in the runner and passed it to the `Scale`.

I might go over this more a few times to try to make this prettier.

Change-Id: I5569c6dfb9b7956cd3e0b3fe3ed5503af84c569e