Add support for decoding Revision from Host. (#10647)
* Add support for decoding Revision from Host.

This adds logic to the activator that falls back on decoding the Host header when `Knative-Serving-Revision` etc is unspecified.  With this, Revisions can simply be curled using their K8s service name and properly activate.

Part of the goal of this is to support a more standard datapath for Revisions, simplifying the contract that headers must be specified.  The current path remains the default, but in the future we might consider leveraging the RewriteHost functionality to use this for the main path as well.

One interesting observation is that with this we could conceivably make Configuration and Revision "Addressable" (modulo GC concerns), which feels like the datapath contract is headed in the right direction.

* vagababov nit