Use consistent alias for autoscaler v1alpha1 everywhere (#10666)
Currently we have the same package aliased as: `av1alpha1`, `asv1a1`,
`pav1alpha1`, `pav1a1`, `asv1alpha1`, `autoscaling`, and
`autoscalingv1alpha1`. I picked `autoscalingv1alpha1` to be consistent.
It's a bit long, but it's the concatenation of last the two packages in
the path so at least there's some logic to it.