Avoid adding gzip-encoding to requests implicitly
Go helpfully adds an Accept-Encoding: gzip header to requests by
default, and silently uncompresses the response if it did so and the
server added compression in response. While this is normally a good
thing, in our case this means we silently add an extra "Accept-Encoding"
header to requests before they get to the user container, and if the
user container compresses the request in response we then pointlessly
uncompress the response inside Queue Proxy. This commit picks up the new
NewProxyAutoTransport method from pkg to avoid this behaviour.