Only log about service scraping on first transition.
When running KNative with Istio, the autoscaler runs in service scraping
mode. Once the autoscaler detects that it should be running in this
mode, it properly latches, and remains there, but it still prints out a
log statement for each scrape attempt. This generates a fairly
substantial volume of logs when there are a number of services active in
a cluster. This changes to only printing out the log statement on
transition rather than every time.