Refactor scrape client to actually take an HTTP request
I pondered ditching the entire interface, seeing as it is not really helping to be honest. We've already passed in a URL with a path and all, so we might as well pass an entire HTTP request.

This makes it easier to muck with said HTTP request in special cases (like adding a header just for direct-pod-scraping).