Update docs for revision timeoutSeconds field (#10806)
Updated the doc comment for timeoutSeconds to align better with what's
in the API spec, but clarified "progress" here actually means "begin
replying".

I think we may also want to introduce a MaxDurationSeconds to do what
TimeoutSeconds sounds like it might do, and at one point did, but I'll
add an issue for that separately.