Be more discerning about which errors cause mesh fall-back
Previous to this commit, we would fall back to service scraping if all
pods we tried to scrape errored. In many cases this heuristic is fine,
but with a small number (e.g. one) of pods, it can lead to us falling
back to service scraping unnecesarily. Once we've turned off direct
scraping, we never turn it back on, so this is unfortunate.

This commit tightens the check to ensure that all of the errors we saw
were compatible with having been caused by the mesh being enabled.