Drop unnecessary parsing from/to string in probe (#11224)
We accept QUEUE_SERVING_PORT as a string env var, parse it to an int,
only to immediately turn it back in to a string with Itoa. There's no
need to do this, and any errors in the supplied port (which is not
user-supplied) will be caught by http.NewRequest when parsing the URL
anyway.