Add API doc generation to codegen (#11238)
This generates api docs for api types when we do all the other code
generation. Doing this is part of trying to simplify the docs release
process, and will remove the need to manually generate these when
releasing docs, which has been a source of pain in docs release. Since
we already need to run update-codegen when modifying api types, there
shouldn't be any significant new friction in doing this (on my machine
time to run codegen went from 4m5s to 4m19s, so you get 14 more seconds
to refresh your coffee).

As an additional bonus, this means you can view API docs for a
particular serving version or commit via github in docs/serving-api.md,
which is nice.