Actually use the rate limited queue (#11256)
* Actually use the rate limited queue

Previously we introduced a rate limited queue, but we still called its Add
method rather than AddRateLimited, so the rate limit didn't do anything.

I'm not wild about the time.Sleep in the test here, but given the bug
that crept in I really wanted to test the actual behaviour rather
than faking the queue out.

We still don't properly use the queue for per-item retries so this won't
work with per-item back-off yet, but at least this rate limits properly.

* Fix my vim being weird with auto-imports today