Use non-pointer workItem in digest resolution (#11273)
* Use non-pointer workItem in digest resolution

This does not change any behaviour since we still immediately Forget
after processing the workItem, but sets the scene for the next PR which
will change that.

With a pointer-based workItem no per-item rate limit will
work since similar items will not be Equal to each other from the
queue's point of view.

* Use read lock

* Defensive nil check