Make digest resolution idempotent
Previously we determined when we had resolved all the digests by
decrementing a `remaining` counter. While this works in the vast
majority of cases, there's a small race where a previous resolve which
was added to the queue starts after Clear. In this case we can have two
resolves for the same image complete, and inaccurately think we've
finished downloading all the requested digests.

This commit uses a map rather than array to collect the results so that
we can ignore duplicate resolves of the same digest.