Do not call GetIngressEndpoint() when --resolvabledomain is enabled
`GetIngressEndpoint()` is a function to get ingress endpoint for the
spoof client. So we don't need to call it when `--resolvabledomain` is
enabled. Especially when Ingresss service uses `ClusterIP`, it alwasy
fails to get endpoint and test fails.

This patch changes to call GetIngressEndpoint() only when
resolvabledomain is false in websocket test.