Unflake per-item test again
The rate limit applies to each image and we fail as soon as either image fails. In
rare cases this could mean that a different image fails on each iteration of the loop,
meaning we don't back off as much as we expect. To simplify this, this commit changes
the test so the same image fails each time.