Add delay for TCP probe test (#11459)
Delays listening on the TCP socket so that if QP goes ready before the
user container is up the test will properly fail. This uses a different
delay from the other cases because we want to ensure that the HTTP and
Exec probe properly check via the /healthz endpoint rather than just
being able to pass via the default tcp probe.