Add permission to update namespaces/finalizers (#11517)
* Add permission to update namespace finalizer

This patch adds the permission to update `namespaces/finalizers`.

Since knative/pkg#2098 added ownerRef refers to namespace for webhook,
we need the permission. Without it, cluster which has a stricter RBAC
rules gets the following error:

```
cannot set blockOwnerDeletion if an ownerReference refers to a resource you can't set finalizers on:
```

* Add kapp workaround

* Add comment

* Remove white space

* Move kapp config into test/config/ytt/core

* remove unnecessary change