Bump to Go 1.16 to take advantage of better FlushInterval behaviour
Prior to Go 1.16 not setting FlushInterval led to streaming responses
never being flushed. To avoid this, we've been setting a low (20ms)
FlushInterval on everything. However as of 1.16 the FlushInterval is
ignored and responses are immediately flushed if the request has a -1
Content-Length, meaning we should be able to set a higher (or no)
FlushInterval without breaking websockets.