Drop unused flag.Parse and simplify emptyDir image even more (#11855)
The flag.Parse was never used because we don't read any flags in this
image so might as well drop it, and the import.

While I was at it I thought we might as well move the handler inline too
since this way we don't need a method call for testfilePath.