Improve error messages on probe failure
Was trying to track down an issue and the current error message is
pretty useless since the first failing verifier is used which means we
just see `unexpected body, want "queue", got ""`. Seeing which status
code we actually got would be much better in most cases where the status
isn't 200.