roles: Don't specify rules when using aggregationRules. (#11899)
* roles: Don't specify rules when using aggregationRules.

According to the Kubernetes API specification:

> If AggregationRule is set, then the Rules are controller managed and
direct changes to Rules will be stomped by the controller.

This changes the configuration to be in line with the documentation. The
issue we ran into was when using FluxCD to manage our clusters, it was
constantly trying to reconcile the `knative-serving-admin` ClusterRole
via a `kubectl apply`. Because the list of roles was specified in the
input yaml, the apply was clearing out the roles list every so often. In
the brief window between the apply clearing the roles, and the cluster
re-creating them we would have spurious permissions errors. Once we
changed the config to no longer specify the rules, as the documentation
suggests, the permissions errors went away and things started to
function better as a result.

* roles: Remove line rather than commenting out.