Speed up concurrency state implementation (#11909)
* Speed up concurrency state implementation

This rewrites the ConcurrencyStateHandler from a channel based approach
into an atomic/mux based approach, to minimize contention and allow for
more concurrent processing.

An RWMutex is used to make sure that calling either pause or resume is
mutually exclusive to each other and also mutually exclusive to the
requests themselves, though the requests can happily run in parallel
(hence them obtaining a read lock).

An important race to consider is a request entering into the pausing
logic while another request enters the resume block. Before pausing,
it's important to recheck if `inFlight` is still 0 **after** obtaining
the lock to make sure no other request has raced obtaining the lock. The
race can happen because the increment/decrement of `inFlight` is not
atomic to getting the lock so the following order of operations could
happen:

1. A: inFlight.Dec() == 0 --> true
2. B: inFlight.Inc() == 1 --> true
3. B: mux.Lock()
4. B: paused --> false
5. B: mux.Unlock()
6. A: mux.Lock()
7. A: !paused --> true
8. A: pause()
9. A: mux.Unlock()
10. B: mux.RLock()
11. B: h.ServeHTTP --> Boom.

In step 7, the pause would not have been excecuted in the actual
implementation though, because we doublecheck the state of `inFlight`,
which wouldn't have been 0 anymore.

* Add waiter mechanism for racing resumes

* Redo things with doubly checked locking and drop waiter channel

* Drop polling for state