Teach activator about mesh compatibility modes (#11976)
This allows us to tell activator explicitly whether mesh is enabled or
disabled, allowing us to skip the heuristic fallbacks when we know mesh
is disabled, and skip the pod probing when we know it is enabled. The
default automatic fallback behaviour remains the default, and this
commit does not actually make this field configurable by a user, yet.