Set old revision defaults in validation
When a revision is updated, revision validation logic ensures revision with the same name is immutable.
This can cause validation error during upgrade when we have new fields in revision spec because validation
considers adding new fields as a mutation. We can solve this by giving new fields in the old revision defaults