Trust k8s readiness signal (#12086)
* Trust k8s readiness signal

If k8s tells us a pod is ready before we manage to get a result via
probing, it makes sense to trust that rather than waiting for a probe
result. This only works when mesh autodetection is off, because
otherwise we need the probe to sniff whether mesh is on.

* Save an allocation