Change default target burst capacity
The current value takes the activator out of the path quite
aggressively, and with min-scale=2 a single request can cause the
activator to flip in and out of the path. Since moving activator in and
out of the path can cause significant load on Istio (etc) this is not
ideal.

This commit defaults TBC to -1 instead so that people can explicitly opt
in to setting TBC based on their actual use case and measured latency
problems caused by having activator in the path that they want to
resolve.

Having activator always in the path does add a small amount of latency
and might increase the number of activators needed in the system, but
empirically our quite large environments have not seen a problem with
this setting.