Disable activator probe optimisation when exec probe is specified
The activator probe relies on Queue Proxy being able to perform the
required probes, but in the case of exec probes it can't. This means
that activator can start proxying to pods before they're ready if an
exec probe is used. This commit disables the optimisation and relies on
the kubernetes readiness signal instead if an exec probe is used.

Note: because we also use the probe for autodetecting mesh mode (unless
explicitly told whether mesh mode is on in the config) we still do probe
even if an exec probe is passed, we just _also_ require kubernetes to
report ready.