Move liveness checks to user container port (#12498)
Otherwise these get intercepted by the drainer logic in queue proxy, which is
not what we want.

Initially liveness and readiness were routed via the queue proxy so that a
single check could check both user container and queue proxy; while this makes
some sense for Readiness (because we need this for optimisations in activator
that need to check readiness of the whole pod at once), for liveness this is
both not needed and actively harmful. A failing liveness probe results in
Kubernetes restarting the failing container (in this case, the user container),
but if the failure comes from QP this will not help (and is actively
unhelpful).  TL;dr: If we want to add a liveness probe to Queue Proxy, we should
add that directly rather than modifying the user's probe to go via the Queue
Proxy port.

Co-authored-by: Julian Friedman <julz.friedman@uk.ibm.com>