Don't scale down istiod during upgrade
istiod Deployment has a initial replica count of 1
istiod's HPA has a minReplica count of 3

During upgrade/downgrade this diff is reconciled by kapp and
istiod deployment is rolled. I believe this is causing our SLI to drop.

This kapp config change will ignore the replica count in the yaml and
opt to select the server's count - thus circumventing the scale down.