Updates val webhook selector to reduce invocations
When we install networking with serving, multiple validating
webhooks are registered.

Previously, the webhooks were configured with just namespace
selectors that are identical. Thus when the config file is updated,
the serving webhook is invoked (which is a noop) and vice-versa
for any serving config maps.

To avoid this, this PR updates the validating webhook to use an
objectSelector and includes an extra label (app.kubernetes.io/component)
to further distinguish net-istio resources from serving resources.

It also adds component labels to several of the configmaps that did
not have them.