[release-1.2] Drop MaxDurationSeconds from the RevisionSpec  (#12640)
* Drop MaxDurationSeconds from the RevisionSpec (#12635)

We added MaxDurationSeconds (#12322) because the behaviour of
RevisionSpec.Timeout changed from total duration to time to first byte.

In hindsight changing the behaviour of Timeout was a mistake since
it goes against the original specification.

Thus we're going to create a path for migration and the first part is
to remove MaxDurationSeconds from the RevisionSpec.

* fix conformance test