adds progressDeadline annotation on per revision basis
Signed-off-by: Paul S. Schweigert <paulschw@us.ibm.com>

This PR allows setting the progress-deadline on each revision by passing
an annotation, autoscaling.knative.dev/progressDeadline. It aims to
improve the experience where revisions don't fail until the
progressDeadline is reached, which by default is 10 minutes. Not
everyone necessarily wants to wait that long, but previously the only
resource was for the cluster admin to lower the deadline
universally (i.e. change the default). With this change, we allow users
to "fail fast", if they so desire.