Fix kpa.TestReconcileDeciderCreatesAndDeletes flake
The test is a bit strict and cannot observe any deletions. But there's a
race when the cache receives the object and the key is enqueued on our
controller.

Thus there are instances where reconciliation occurs but the item is not
present in the cache. Resulting in a failure.

To mitigate this we now wait to start the controller until the desired
object is in the informer's cache.