ServerlessService's initial operation mode is now Proxy (#12842)
Prior the initial operation mode was Serve.  This would create a
narrow window where the revision's pods directly receive requests
from the ingress.  The size of this window depends on when the
Revision's Pod becomes ready and the time it takes the autoscaler
to make an initial decision on scale an excess burst capacity.

If there is a continuous traffic this could overwhelm Revisions
with a low container concurrency causing the queue-proxy to return
503s since it's queue is full.