modify checksum so its not a float64
Signed-off-by: Paul S. Schweigert <paulschw@us.ibm.com>

This PR introduces a trivial change to the example block to update the
generated checksum.

The previous value was 45463e45, which if not quoted will render as a
float64. This caused problems when applied via kapp through the CI,
because the annotation needs to be of type string, even if not
quoted (which was what was happening with kapp).

This new checksum, since it starts with "aa", shouldn't have that
problem.

Given the incredibly low odds of this having happened in the first
place, I'm off to buy a lottery ticket now... :)