Improve SKS handling for unavailable Activator.
This change refactors the logic that handles the SKS "mode" a bit.
Currently, there is logic when the SKS is in Serve mode to continue
using the activator subset when there aren't ready private endpoints.
This adds similar logic to the activator side, but also restructures
how this is done to orient it around forcing the mode one way or the
other.

With this change, if there are no available activator endpoints, then
we will force "Serve" mode.  If there are available activator endpoints
but not available revision endpoints, then we will force "Proxy" mode.

I've hoisted some of the (mostly debug) logging we do, and simplified
the switch statement to simply key directly off the mode that results
from the above criteria.