Remove the scale-to-zero checking in conformance. (#13025)
Conformance should be a measurement of the Knative API and runtime semantics, not the details of how those semantics are implemented.  Scaling to zero is effectively a cost optimization, but should not be a requirement of a
conforming implementation.

For instance, if I were to set `min-scale: "1"` cluster-wide, I should still have a conformant implementation, since the key semantics of the API are still preserved.