Relax the "timeout" requirement for `TestProbeRuntimeAfterStartup` (#13026)
Currently the `TestProbeRuntimeAfterStartup` conformance test requires
that implementations match exactly the semantics of the Knative default
configuration, which feels disconnected from the purpose of conformance.

At an example, if I configure Knative slightly differently, so that the
activator isn't in the request path (TBC=0), then the failure mode of
this test changes from timing out to serving 503s (what K8s would do).

This touches on the point Julz was making in #10765 where the current
behavior of timing out is an unfortunate departure from K8s semantics,
but only when TBC>0.

This change allows 503s as an acceptable failure mode of this conformance
test.