Temporarily disable flaky kind test (#13055)
* temp. disable flaky kind test

Signed-off-by: Paul S. Schweigert <paulschw@us.ibm.com>

TestAutoscaleSustaining/aggregation-weighted-exponential is currently
extremely flaky when running on Kind. There were some major updates
made in the recent version of kind, so will need to investigate
further to determine what changed and why things are so flaky. In the
interim, disabling this test so as not to spam the grid with false
negatives (as the test passes fine when running on GKE).

See #13049 for tracking of the Kind investigation.

* skip both tests

Signed-off-by: Paul S. Schweigert <paulschw@us.ibm.com>