Add active-min-scale annotation (#13136)
* add min-non-zero-replicas option for scaling 0 => n, n>1

Signed-off-by: Paul S. Schweigert <paulschw@us.ibm.com>

This PR implements a "min-scale, but with scale to zero" feature. In
essence, when this feature is in use, if the service is handling
requests it will never have less than N number of pods (though it can
still scale to zero).

N is specified on a per-revision basis using an
`autoscaling.knative.dev/min-non-zero-replicas` annotation.

If min-non-zero-replicas > initial-scale, the service will initially
scale to the min-non-zero-replicas value.

* review fixes

Signed-off-by: Paul S. Schweigert <paulschw@us.ibm.com>

* fix unit test

Signed-off-by: Paul S. Schweigert <paulschw@us.ibm.com>

* renaming to active-min-scale

Signed-off-by: Paul S. Schweigert <paulschw@us.ibm.com>