ConfigurationStatus.LatesReadyRevision will only advance forward
The search for latest ready revision (LLR) looked at each revision's
configurationGeneration label to search for a new candidate.

Only Revisions with a configurationGeneration in a certain range were
considered. Prior this inclusive range was [LLR.generation, config.generation].

The lower bound was incorrect and it should actually be LLR.configGeneration.
Otherwise LLR generation is fairly low the the set of revision to consider
would be quite large for really old Configurations.