Fix LatestReadyRevision semantics - it only advances forward (#13239)
* Configuration traffic targets now check Revision readiness

* ConfigurationStatus.LatesReadyRevision will only advance forward

The search for latest ready revision (LLR) looked at each revision's
configurationGeneration label to search for a new candidate.

Only Revisions with a configurationGeneration in a certain range were
considered. Prior this inclusive range was [LLR.generation, config.generation].

The lower bound was incorrect and it should actually be LLR.configGeneration.
Otherwise LLR generation is fairly low the the set of revision to consider
would be quite large for really old Configurations.

* Using BYO revision name will allow latest created/ready to rollback.

This is required to support git revert in combination with a Knative
Service that has latestRevision: true in their Route traffic block

* Include BYOName Revision Generation

Prior the conditional was not correct and the generation was not part of
the selector. Also made sure the unit test failed prior to applying the
fix.